refactor change password and signup
