--- apps/doji-backend/src/register/register.module.ts
@@ -1,14 +1,0 @@
-import { User } from '@backend/entities/User'
-import { MikroOrmModule } from '@mikro-orm/nestjs'
-import { Module } from '@nestjs/common'
-
-import { RegisterController } from './register.controller'
-import { RegisterService } from './register.service'
-
-@Module({
-  imports: [MikroOrmModule.forFeature([User])],
-  controllers: [RegisterController],
-  providers: [RegisterService],
-  exports: [RegisterService],
-})
-export class RegisterModule {}
