+++ libs/api/src/index.ts
@@ -1,7 +1,6 @@
 export * from './lib/api'
 export * from './lib/messaging'
 export * from './lib/dtos/auth'
-export * from './lib/dtos/register'
 export * from './lib/dtos/payment'
 export * from './lib/dtos/expertApp'
 export * from './lib/entities/User'
