--- libs/api/src/lib/api.spec.ts
@@ -1,7 +1,0 @@
-import { api } from './api'
-
-describe('api', () => {
-  it('should work', () => {
-    expect(api()).toEqual('api')
-  })
-})
