--- libs/api/src/lib/dtos/register.ts
@@ -1,16 +1,0 @@
-export interface IUserChangePasswordRequestDTO {
-  currentPassword: string
-  newPassword: string
-}
-
-export interface IUserRegistrationRequestDTO {
-  username: string
-  displayName: string
-  email: string
-  password: string
-}
-
-export interface IAdminCreationRequestDTO {
-  username: string
-  password: string
-}
